requirements.txt
Django==3.0.8
django-extensions==3.0.3
django-username-email==2.3.0
djangorestframework==3.11.0
flake8==3.8.3
